Development

+++ models/model/membership_package.go
@@ -7,12 +7,12 @@ import (
 )
 
 type MembershipPackage struct {
-	ID 			uint64 `gorm:"primaryKey;notNull"`
-	AdminID		uint64
-	Name 		string
-	Duration 	uint16
-	CreatedAt 	time.Time
-	UpdatedAt 	time.Time
-	DeletedAt 	gorm.DeletedAt
+	ID                     uint64 `gorm:"primaryKey;notNull"`
+	AdminID                uint64
+	Name                   string
+	Duration               uint32
+	CreatedAt              time.Time
+	UpdatedAt              time.Time
+	DeletedAt              gorm.DeletedAt
 	MembershipTransactions []MembershipTransaction
 }
